debthound/requirements.txt
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
tox
appdirs==1.4.3
requests
scrapyd==1.2.0
schedule-0.5.0